feat(autopairs): Move nvim-autopairs to coding.
