+++ lua/core/utils/lazyLoad.lua
@@ -6,22 +6,23 @@ return function(plugin)
       local condition = file ~= "NvimTree_1" and file ~= "[lazy]" and file ~= ""
 
       if condition then
-	vim.api.nvim_del_augroup_by_name("BeLazyOnFileOpen" .. plugin)
+        vim.api.nvim_del_augroup_by_name("BeLazyOnFileOpen" .. plugin)
 
-	-- dont defer for treesitter as it will show slow highlighting
-	-- This deferring only happens only when we do "nvim filename"
-	if plugin ~= "nvim-treesitter" then
-	  vim.schedule(function()
-	    require("lazy").load { plugins = plugin }
+        -- dont defer for treesitter as it will show slow highlighting
+        -- This deferring only happens only when we do "nvim filename"
+        if plugin ~= "nvim-treesitter" then
+          vim.schedule(function()
+            require("lazy").load { plugins = plugin }
 
-	    if plugin == "nvim-lspconfig" then
-	      vim.cmd "silent! do FileType"
-	    end
-	  end, 0)
-	else
-	  require("lazy").load { plugins = plugin }
-	end
+            if plugin == "nvim-lspconfig" then
+              vim.cmd "silent! do FileType"
+            end
+          end, 0)
+        else
+          require("lazy").load { plugins = plugin }
+        end
       end
     end,
   })
 end
+
