+++ lua/plugins/configs/editor/treesitter.lua
@@ -7,9 +7,5 @@ require("nvim-treesitter.configs").setup({
     enable = true,
     use_language_tree = true,
   },
-
-  incremental_selection = {
-    enable = false,
-  },
 })
